requirements.txt
djangorestframework==3.3.1
wheel==0.24.0
